Convert EtcdCentralStoreThread to a util
