--- tendrl/alerting/central_store/exceptions.py
@@ -1,2 +1,0 @@
-class EtcdError(Exception):
-    pass
